% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{IQRange}
\alias{IQRange}
\title{Inter Quartile Range Value}
\usage{
IQRange(dataframe, column)
}
\arguments{
\item{dataframe}{The chosen data frame being used for study}

\item{column}{The chosen column used to analyze and reduce potential outliers}
}
\value{
the Inter Quartile Range Value
}
\description{
Inter Quartile Range Value
}
\examples{
IQRange(mtcars, 'mpg')
}
