% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{maxQuantileValue}
\alias{maxQuantileValue}
\title{Maximum Quantile Value}
\usage{
maxQuantileValue(dataframe, column)
}
\arguments{
\item{dataframe}{The chosen data frame being used for study}

\item{column}{The chosen column used to analyze and reduce potential outliers}
}
\value{
the Quantile value at .75
}
\description{
Maximum Quantile Value
}
\examples{
maxQuantileValue(mtcars, 'mpg')
}
