% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{minQuantileValue}
\alias{minQuantileValue}
\title{Minimum Quantile Value}
\usage{
minQuantileValue(dataframe, column)
}
\arguments{
\item{dataframe}{The chosen data frame being used for study}

\item{column}{The chosen column used to analyze and reduce potential outliers}
}
\value{
the Quantile value at .25
}
\description{
Minimum Quantile Value
}
\examples{
minQuantileValue(mtcars, 'mpg')
}
