% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{quantileRemove}
\alias{quantileRemove}
\title{Remove Outliers using the IQR Method}
\usage{
quantileRemove(dataframe, column)
}
\arguments{
\item{dataframe}{The chosen data frame being used for study}

\item{column}{The chosen column used to analyze and reduce potential outliers}
}
\value{
The dataframe argument with observation outliers from the requested column removed
}
\description{
Remove Outliers using the IQR Method
}
\examples{
quantileRemove(mtcars,'mpg')
}
