% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{zScoreRemove}
\alias{zScoreRemove}
\title{Remove the Outliers using the Z-score method}
\usage{
zScoreRemove(dataframe, column)
}
\arguments{
\item{dataframe}{The chosen data frame being used for study}

\item{column}{The chosen column used to analyze and reduce potential outliers}
}
\value{
A dataframe with observation outliers from the selected column over a z-score of 3 removed
}
\description{
Remove the Outliers using the Z-score method
}
\examples{
zScoreRemove(mtcars, 'mpg')
}
